+++ app/src/main/java/com/alaboud/businesssuite/MainActivity.kt
@@ -103,7 +103,7 @@ class MainActivity : AppCompatActivity() {
             mediaPlaybackRequiresUserGesture = false
             cacheMode = WebSettings.LOAD_DEFAULT
             mixedContentMode = WebSettings.MIXED_CONTENT_NEVER_ALLOW
-            userAgentString = "$userAgentString AlAboudMobile/24.2.4"
+            userAgentString = "$userAgentString AlAboudMobile/25.3.3"
         }
 
         CookieManager.getInstance().apply {